+++ src/trunk/libs/3rd-party/libbson/examples/bcon-col-view.c
@@ -14,8 +14,7 @@
  * limitations under the License.
  */
 
-#include <bson.h>
-#include <bcon.h>
+#include <bson/bson.h>
 
 #define QUERY(...) ELE_QUERY, __VA_ARGS__, NULL
 #define SORT(...) ELE_SORT, __VA_ARGS__, NULL