+++ src/trunk/libs/3rd-party/libbson/examples/bcon-speed.c
@@ -15,8 +15,7 @@
  */
 
 
-#include <bson.h>
-#include <bcon.h>
+#include <bson/bson.h>
 #include <stdio.h>
 #include <stdlib.h>