+++ src/trunk/libs/3rd-party/libbson/src/bson/bson-context.h
@@ -14,18 +14,15 @@
  * limitations under the License.
  */
 
+#include "bson/bson-prelude.h"
+
 
 #ifndef BSON_CONTEXT_H
 #define BSON_CONTEXT_H
 
 
-#if !defined(BSON_INSIDE) && !defined(BSON_COMPILATION)
-#error "Only <bson.h> can be included directly."
-#endif
-
-
-#include "bson-macros.h"
-#include "bson-types.h"
+#include "bson/bson-macros.h"
+#include "bson/bson-types.h"
 
 
 BSON_BEGIN_DECLS
@@ -36,7 +33,7 @@ bson_context_new (bson_context_flags_t flags);
 BSON_EXPORT (void)
 bson_context_destroy (bson_context_t *context);
 BSON_EXPORT (bson_context_t *)
-bson_context_get_default (void) BSON_GNUC_CONST;
+bson_context_get_default (void);
 
 
 BSON_END_DECLS