+++ src/trunk/libs/3rd-party/libbson/src/bson/bson-memory.c
@@ -17,10 +17,12 @@
 
 #include <stdlib.h>
 #include <string.h>
+#include <stdio.h>
+#include <errno.h>
 
-#include "bson-atomic.h"
-#include "bson-config.h"
-#include "bson-memory.h"
+#include "bson/bson-atomic.h"
+#include "bson/bson-config.h"
+#include "bson/bson-memory.h"
 
 
 static bson_mem_vtable_t gMemVtable = {
@@ -66,6 +68,7 @@ bson_malloc (size_t num_bytes) /* IN */
 
    if (BSON_LIKELY (num_bytes)) {
       if (BSON_UNLIKELY (!(mem = gMemVtable.malloc (num_bytes)))) {
+         fprintf (stderr, "Failure to allocate memory in bson_malloc(). errno: %d.\n", errno);
          abort ();
       }
    }
@@ -103,6 +106,7 @@ bson_malloc0 (size_t num_bytes) /* IN */
 
    if (BSON_LIKELY (num_bytes)) {
       if (BSON_UNLIKELY (!(mem = gMemVtable.calloc (1, num_bytes)))) {
+         fprintf (stderr, "Failure to allocate memory in bson_malloc0(). errno: %d.\n", errno);
          abort ();
       }
    }
@@ -150,6 +154,7 @@ bson_realloc (void *mem,        /* IN */
    mem = gMemVtable.realloc (mem, num_bytes);
 
    if (BSON_UNLIKELY (!mem)) {
+      fprintf (stderr, "Failure to re-allocate memory in bson_realloc(). errno: %d.\n", errno);
       abort ();
    }
 
@@ -257,9 +262,9 @@ bson_zero_free (void *mem,   /* IN */
  *
  * bson_mem_set_vtable --
  *
- *       This function will change our allocationt vtable.
+ *       This function will change our allocation vtable.
  *
- *       It is imperitive that this is called at the beginning of the
+ *       It is imperative that this is called at the beginning of the
  *       process before any memory has been allocated by the default
  *       allocator.
  *