+++ src/trunk/libs/3rd-party/libbson/src/bson/bson-private.h
@@ -14,14 +14,16 @@
  * limitations under the License.
  */
 
+#include "bson/bson-prelude.h"
+
 
 #ifndef BSON_PRIVATE_H
 #define BSON_PRIVATE_H
 
 
-#include "bson-macros.h"
-#include "bson-memory.h"
-#include "bson-types.h"
+#include "bson/bson-macros.h"
+#include "bson/bson-memory.h"
+#include "bson/bson-types.h"
 
 
 #if (__GNUC__ > 4) || (__GNUC__ == 4 && __GNUC_MINOR__ >= 6)
@@ -54,17 +56,25 @@ typedef enum {
 } bson_flags_t;
 
 
+#ifdef BSON_MEMCHECK
+#define BSON_INLINE_DATA_SIZE (120 - sizeof (char*))
+#else
 #define BSON_INLINE_DATA_SIZE 120
+#endif
+
 
 BSON_ALIGNED_BEGIN (128)
 typedef struct {
    bson_flags_t flags;
    uint32_t len;
+#ifdef BSON_MEMCHECK
+   char *canary;
+#endif
    uint8_t data[BSON_INLINE_DATA_SIZE];
 } bson_impl_inline_t BSON_ALIGNED_END (128);
 
 
-BSON_STATIC_ASSERT (sizeof (bson_impl_inline_t) == 128);
+BSON_STATIC_ASSERT2 (impl_inline_t, sizeof (bson_impl_inline_t) == 128);
 
 
 BSON_ALIGNED_BEGIN (128)
@@ -83,7 +93,7 @@ typedef struct {
 } bson_impl_alloc_t BSON_ALIGNED_END (128);
 
 
-BSON_STATIC_ASSERT (sizeof (bson_impl_alloc_t) <= 128);
+BSON_STATIC_ASSERT2 (impl_alloc_t, sizeof (bson_impl_alloc_t) <= 128);
 
 
 #define BSON_REGEX_OPTIONS_SORTED "ilmsux"