+++ src/trunk/libs/3rd-party/libbson/src/bson/bson-reader.c
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-#include "bson.h"
+#include "bson/bson.h"
 
 #include <errno.h>
 #include <fcntl.h>
@@ -27,8 +27,8 @@
 #include <sys/stat.h>
 #include <sys/types.h>
 
-#include "bson-reader.h"
-#include "bson-memory.h"
+#include "bson/bson-reader.h"
+#include "bson/bson-memory.h"
 
 
 typedef enum {
@@ -644,7 +644,9 @@ _bson_reader_data_tell (bson_reader_data_t *reader) /* IN */
 void
 bson_reader_destroy (bson_reader_t *reader) /* IN */
 {
-   BSON_ASSERT (reader);
+   if (!reader) {
+      return;
+   }
 
    switch (reader->type) {
    case 0: