+++ CHANGELOG.md
@@ -1,5 +1,46 @@
 # Jakarta
 
+## Release YYYY.ddd
+
+* trunk
+
+  * Add distance and depth range confguration to amplitude profiles
+    in scconfig
+  * Fix PostgreSQL database plugin to unescape bytea encoded data
+
+* scmergexml
+
+  * New module: Merge the content of multiple XML files in SC3ML format. Currently only
+    event parameters are supported.
+
+* scmag
+
+  * Add ```--keep-weights``` flag to retain station magnitude weights
+    when reprocessing with ```--static```
+
+* scautoloc
+
+  * Fix station lookup with inventories with overlapping network epochs
+  * Do not use database if inventory should be read from a file
+
+* fdsnws
+
+  * Fix dataselect restricted handling with respect to station service
+
+    * Require valid user if network or station is restricted
+    * Stop iteration at network or station level if restricted flag but
+      no user is present 
+
+* fdsnws
+
+  * Fix dataselect restricted handling with respect to station service
+
+    * Require valid user if network or station is restricted
+    * Stop iteration at network or station level if restricted flag but
+      no user is present
+
+  * Implement service specific version numbers 
+
 ## Release 2018.327 patch12
 
 * scamp